+++ apps/widget/modules/widget/ui/screens/widget-chat-screen/index.tsx
@@ -1,6 +1,7 @@
 "use client"
 
 import { z } from "zod"
+import { useMemo } from "react"
 import { useForm } from "react-hook-form"
 import { useAction, useQuery } from "convex/react"
 import { useAtomValue, useSetAtom } from "jotai"
@@ -21,6 +22,7 @@ import {
   conversationIdAtom,
   organizationIdAtom,
   screenAtom,
+  widgetSettingsAtom,
 } from "@/modules/widget/atoms/widget-atoms"
 
 import { Button } from "@workspace/ui/components/button"
@@ -54,10 +56,10 @@ const formSchema = z.object({
 
 export const WidgetChatScreen = () => {
   const setScreen = useSetAtom(screenAtom)
-  const setConversationId = useSetAtom(conversationIdAtom)
-
+  const widgetSettings = useAtomValue(widgetSettingsAtom)
   const conversationId = useAtomValue(conversationIdAtom)
   const organizationId = useAtomValue(organizationIdAtom)
+  const setConversationId = useSetAtom(conversationIdAtom)
   const contactSessionId = useAtomValue(
     contactSessionIdAtomFamily(organizationId || "")
   )
@@ -66,6 +68,18 @@ export const WidgetChatScreen = () => {
     setConversationId(null)
     setScreen("selection")
   }
+
+  const suggestions = useMemo(() => {
+    if (!widgetSettings) {
+      return []
+    }
+
+    return Object.keys(widgetSettings.defaultSuggestions).map((key) => {
+      return widgetSettings.defaultSuggestions[
+        key as keyof typeof widgetSettings.defaultSuggestions
+      ]
+    })
+  }, [widgetSettings])
 
   const conversation = useQuery(
     api.public.conversations.getOne,
@@ -160,7 +174,30 @@ export const WidgetChatScreen = () => {
           })}
         </AIConversationContent>
       </AIConversation>
-      {/* TODO: Add suggestions */}
+      {toUIMessages(messages.results ?? [])?.length === 1 && (
+        <AISuggestions className="flex w-full flex-col items-end p-2">
+          {suggestions.map((suggestion) => {
+            if (!suggestion) {
+              return null
+            }
+
+            return (
+              <AISuggestion
+                key={suggestion}
+                onClick={() => {
+                  form.setValue("message", suggestion, {
+                    shouldValidate: true,
+                    shouldDirty: true,
+                    shouldTouch: true,
+                  })
+                  form.handleSubmit(onSubmit)()
+                }}
+                suggestion={suggestion}
+              />
+            )
+          })}
+        </AISuggestions>
+      )}
       <Form {...form}>
         <AIInput
           className="rounded-none border-x-0 border-b-0"